move files around and make a chonky migration
